64 bit integer on rationals

+++ extern/zx/include/Rational.hpp
@@ -15,26 +15,26 @@ class Rational {
   void normalize();
 
 public:
-  int32_t num, denom;
+  int64_t num, denom;
 
   Rational() : num(0), denom(1){};
-  explicit Rational(int32_t num, int32_t denom) : num(num), denom(denom) {
+  explicit Rational(int64_t num, int64_t denom) : num(num), denom(denom) {
     normalize();
   }
-  explicit Rational(int32_t num) : num(num), denom(1) { normalize(); }
+  explicit Rational(int64_t num) : num(num), denom(1) { normalize(); }
   explicit Rational(double val);
 
   Rational &operator+=(const Rational &rhs);
-  Rational &operator+=(const int32_t rhs);
+  Rational &operator+=(const int64_t rhs);
 
   Rational &operator-=(const Rational &rhs);
-  Rational &operator-=(const int32_t rhs);
+  Rational &operator-=(const int64_t rhs);
 
   Rational &operator*=(const Rational &rhs);
-  Rational &operator*=(const int32_t rhs);
+  Rational &operator*=(const int64_t rhs);
 
   Rational &operator/=(const Rational &rhs);
-  Rational &operator/=(const int32_t rhs);
+  Rational &operator/=(const int64_t rhs);
 
   double to_double() const;
   bool is_integer() const { return denom == 1; }
@@ -47,11 +47,11 @@ inline Rational operator+(Rational lhs, const Rational &rhs) {
   lhs += rhs;
   return lhs;
 }
-inline Rational operator+(Rational lhs, const int32_t rhs) {
+inline Rational operator+(Rational lhs, const int64_t rhs) {
   lhs += rhs;
   return lhs;
 }
-inline Rational operator+(const int32_t lhs, Rational rhs) {
+inline Rational operator+(const int64_t lhs, Rational rhs) {
   rhs += lhs;
   return rhs;
 }
@@ -60,11 +60,11 @@ inline Rational operator-(Rational lhs, const Rational &rhs) {
   lhs -= rhs;
   return lhs;
 }
-inline Rational operator-(Rational lhs, const int32_t rhs) {
+inline Rational operator-(Rational lhs, const int64_t rhs) {
   lhs -= rhs;
   return lhs;
 }
-inline Rational operator-(const int32_t lhs, Rational rhs) {
+inline Rational operator-(const int64_t lhs, Rational rhs) {
   rhs -= lhs;
   return rhs;
 }
@@ -73,11 +73,11 @@ inline Rational operator*(Rational lhs, const Rational &rhs) {
   lhs *= rhs;
   return lhs;
 }
-inline Rational operator*(Rational lhs, const int32_t rhs) {
+inline Rational operator*(Rational lhs, const int64_t rhs) {
   lhs *= rhs;
   return lhs;
 }
-inline Rational operator*(const int32_t lhs, Rational rhs) {
+inline Rational operator*(const int64_t lhs, Rational rhs) {
   rhs *= lhs;
   return rhs;
 }
@@ -86,11 +86,11 @@ inline Rational operator/(Rational lhs, const Rational &rhs) {
   lhs /= rhs;
   return lhs;
 }
-inline Rational operator/(Rational lhs, const int32_t rhs) {
+inline Rational operator/(Rational lhs, const int64_t rhs) {
   lhs /= rhs;
   return lhs;
 }
-inline Rational operator/(const int32_t lhs, Rational rhs) {
+inline Rational operator/(const int64_t lhs, Rational rhs) {
   rhs /= lhs;
   return rhs;
 }
@@ -99,11 +99,11 @@ inline bool operator<(const Rational &lhs, const Rational &rhs) {
   return lhs.num * rhs.denom < rhs.num * lhs.denom;
 }
 
-inline bool operator<(const Rational &lhs, int32_t rhs) {
+inline bool operator<(const Rational &lhs, int64_t rhs) {
   return lhs.num < rhs * lhs.denom;
 }
 
-inline bool operator<(int32_t lhs, const Rational &rhs) {
+inline bool operator<(int64_t lhs, const Rational &rhs) {
   return lhs * rhs.denom < rhs.num;
 }
 
@@ -111,11 +111,11 @@ inline bool operator<=(const Rational &lhs, const Rational &rhs) {
   return lhs.num * rhs.denom <= rhs.num * lhs.denom;
 }
 
-inline bool operator<=(const Rational &lhs, int32_t rhs) {
+inline bool operator<=(const Rational &lhs, int64_t rhs) {
   return lhs.num <= rhs * lhs.denom;
 }
 
-inline bool operator<=(int32_t lhs, const Rational &rhs) {
+inline bool operator<=(int64_t lhs, const Rational &rhs) {
   return lhs * rhs.denom <= rhs.num;
 }
 
@@ -123,37 +123,37 @@ inline bool operator>(const Rational &lhs, const Rational &rhs) {
   return rhs < lhs;
 }
 
-inline bool operator>(const Rational &lhs, int32_t rhs) { return rhs < lhs; }
+inline bool operator>(const Rational &lhs, int64_t rhs) { return rhs < lhs; }
 
-inline bool operator>(int32_t lhs, const Rational &rhs) { return rhs < lhs; }
+inline bool operator>(int64_t lhs, const Rational &rhs) { return rhs < lhs; }
 
 inline bool operator>=(const Rational &lhs, const Rational &rhs) {
   return rhs <= lhs;
 }
 
-inline bool operator>=(const Rational &lhs, int32_t rhs) { return rhs <= lhs; }
+inline bool operator>=(const Rational &lhs, int64_t rhs) { return rhs <= lhs; }
 
-inline bool operator>=(int32_t lhs, const Rational &rhs) { return rhs <= lhs; }
+inline bool operator>=(int64_t lhs, const Rational &rhs) { return rhs <= lhs; }
 
 inline bool operator==(const Rational &lhs, const Rational &rhs) {
   return lhs.num == rhs.num && lhs.denom == rhs.denom;
 }
 
-inline bool operator==(const Rational &lhs, int32_t rhs) {
+inline bool operator==(const Rational &lhs, int64_t rhs) {
   return lhs.num == rhs && lhs.denom == 1;
 }
 
-inline bool operator==(int32_t lhs, const Rational &rhs) { return rhs == lhs; }
+inline bool operator==(int64_t lhs, const Rational &rhs) { return rhs == lhs; }
 
 inline bool operator!=(const Rational &lhs, const Rational &rhs) {
   return !(lhs == rhs);
 }
 
-inline bool operator!=(const Rational &lhs, int32_t rhs) {
+inline bool operator!=(const Rational &lhs, int64_t rhs) {
   return !(lhs == rhs);
 }
 
-inline bool operator!=(int32_t lhs, const Rational &rhs) {
+inline bool operator!=(int64_t lhs, const Rational &rhs) {
   return !(lhs == rhs);
 }
